requirements.txt
chex
dm-haiku
# evosax
flax
# jax
# jaxopt
matplotlib
ml-collections
numpy
optax
plotly
tqdm
wandb